requirements.txt
gunicorn
django-heroku
dj-database-url
whitenoise
psycopg2-binary
